+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 # [Unreleased]
+### Added
+- Added UMM option to change the default clothes on the Isekai Protagonist Class.
 ### Changed
 - Changed `Isekai Protagonist`'s `Friendly Aura` obtained from level 9 -> 10.
 - Changed `Isekai Protagonist` and `Villain`'s '`Quicked-Footed` obtained from level 15 -> 16.